requirements.txt
ccxt
yfinance
pandas_ta
python-dotenv
requests
alpaca-py
pandas_market_calendars
pexpect
termcolor